+++ x/exchange/export_test.go
@@ -3,6 +3,4 @@ package exchange
 var (
 	// CreatePaymentGetSignersFunc is a test-only exposure of createPaymentGetSignersFunc.
 	CreatePaymentGetSignersFunc = createPaymentGetSignersFunc
-	// AllRequestMsgs is a test-only exposure of allRequestMsgs.
-	AllRequestMsgs = allRequestMsgs
 )