+++ x/oracle/types/codec.go
@@ -1,27 +1,19 @@
 package types
 
 import (
-	"github.com/cosmos/cosmos-sdk/codec"
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/gogoproto/proto"
 )
 
-// ignoring RegisterLegacyAminoCodec registers all the necessary types and interfaces for the
-// double check
+// RegisterInterfaces registers concrete implementations for this module.
 func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
-	registry.RegisterImplementations(
-		(*sdk.Msg)(nil),
-		&MsgUpdateOracleRequest{},
-		&MsgSendQueryOracleRequest{},
-	)
+	messages := make([]proto.Message, len(AllRequestMsgs))
+	copy(messages, AllRequestMsgs)
+	registry.RegisterImplementations((*sdk.Msg)(nil), messages...)
 
 	registry.RegisterImplementations((*proto.Message)(nil),
 		&QueryOracleRequest{},
 		&QueryOracleResponse{},
 	)
 }
-
-var (
-	ModuleCdc = codec.NewProtoCodec(cdctypes.NewInterfaceRegistry())
-)